requirements.txt
sphinx>=5.0.0
myst-parser>=0.18.0
beautifulsoup4>=4.11.1
pandas>=2.2.3
selenium>=4.30.0
Pillow==11.1.0
psycopg_binary>=3.2.6
psycopg>=3.2.6
python-dotenv==1.0.1
pyodbc==5.2.0
pptx2md==1.7.1
python-pptx==1.0.2
openai==1.100.2
tiktoken==0.11.0
langfuse==3.7.0
transformers==4.55.3
torch==2.8.0
scikit-learn==1.6.1
matplotlib==3.9.4